+++ CHANGELOG.md
@@ -1,5 +1,65 @@
 # Changelog
 
+## [4.3.5](https://github.com/zenbones-theme/zenbones.nvim/compare/v4.3.4...v4.3.5) (2024-08-24)
+
+
+### Bug Fixes
+
+* add lazy.nvim instructions ([6c01e4c](https://github.com/zenbones-theme/zenbones.nvim/commit/6c01e4cd3a4e8ae4321069bfb814cac1e2b2a07e))
+* improve diff ([168593f](https://github.com/zenbones-theme/zenbones.nvim/commit/168593f4ad8dc79fce52b88b5cf90447536d0a81))
+
+## [4.3.4](https://github.com/zenbones-theme/zenbones.nvim/compare/v4.3.3...v4.3.4) (2024-08-23)
+
+
+### Bug Fixes
+
+* Leap highlight are out of date ([#195](https://github.com/zenbones-theme/zenbones.nvim/issues/195)) ([b433225](https://github.com/zenbones-theme/zenbones.nvim/commit/b433225e29de5955a70b3e25a762cfaf2cf24a3d))
+
+## [4.3.3](https://github.com/zenbones-theme/zenbones.nvim/compare/v4.3.2...v4.3.3) (2024-08-08)
+
+
+### Bug Fixes
+
+* **specs:** use Constant instead of Number ([#192](https://github.com/zenbones-theme/zenbones.nvim/issues/192)) ([a041a4e](https://github.com/zenbones-theme/zenbones.nvim/commit/a041a4e7a10b8f185f3eaf18e35e60cb0d3ca6a2))
+
+## [4.3.2](https://github.com/zenbones-theme/zenbones.nvim/compare/v4.3.1...v4.3.2) (2024-08-08)
+
+
+### Bug Fixes
+
+* **ci:** add tagging step back ([#189](https://github.com/zenbones-theme/zenbones.nvim/issues/189)) ([1acd87e](https://github.com/zenbones-theme/zenbones.nvim/commit/1acd87ea10ce26b1613167b7e060ff7110f20686))
+* **ci:** fix release step ([#185](https://github.com/zenbones-theme/zenbones.nvim/issues/185)) ([b42f399](https://github.com/zenbones-theme/zenbones.nvim/commit/b42f3994a9079c192eba674dd8fb8acbacca80d8))
+* **ci:** proper permissions ([#187](https://github.com/zenbones-theme/zenbones.nvim/issues/187)) ([26e26dc](https://github.com/zenbones-theme/zenbones.nvim/commit/26e26dc704f0ce037fca4b96742f2224292fbe07))
+* **ci:** remove deprecated release ([#186](https://github.com/zenbones-theme/zenbones.nvim/issues/186)) ([65d324d](https://github.com/zenbones-theme/zenbones.nvim/commit/65d324d2d703296b8c9872063a8733ce34d69fd4))
+* **ci:** update ([#188](https://github.com/zenbones-theme/zenbones.nvim/issues/188)) ([9cf9de9](https://github.com/zenbones-theme/zenbones.nvim/commit/9cf9de96db145854927ac4621772b7161d38935c))
+* **ci:** update auto commit ([#184](https://github.com/zenbones-theme/zenbones.nvim/issues/184)) ([ce694ef](https://github.com/zenbones-theme/zenbones.nvim/commit/ce694ef8e222e4682ebd9c7de62cf2747efccea0))
+* **doc:** update install ([e9ba830](https://github.com/zenbones-theme/zenbones.nvim/commit/e9ba8305f65524f9355515cdd8704b2c6ba2def6))
+* **specs:** sync light and dark ([c4e1845](https://github.com/zenbones-theme/zenbones.nvim/commit/c4e1845cdc9be2c1b2ba02c78ee0438e576898cf))
+* **ts:** use delimiter for [@constructor](https://github.com/constructor).lua ([1ac520f](https://github.com/zenbones-theme/zenbones.nvim/commit/1ac520f7fc8b1c41d1414270ee48c8f8be2714b0))
+
+## [4.3.1](https://github.com/mcchrish/zenbones.nvim/compare/v4.3.0...v4.3.1) (2024-05-20)
+
+
+### Bug Fixes
+
+* lsp inlay hint ([#180](https://github.com/mcchrish/zenbones.nvim/issues/180)) ([0554b55](https://github.com/mcchrish/zenbones.nvim/commit/0554b55153eaea443402e282e0064eeb7a0f75a7))
+
+## [4.3.0](https://github.com/mcchrish/zenbones.nvim/compare/v4.2.0...v4.3.0) (2024-05-19)
+
+
+### Features
+
+* **ci:** sync to codeberg and gitlab ([#168](https://github.com/mcchrish/zenbones.nvim/issues/168)) ([f5f5fff](https://github.com/mcchrish/zenbones.nvim/commit/f5f5fff4e9ebb68c956a4557790a8941fbcf9c1d))
+* support fzf-lua ([#178](https://github.com/mcchrish/zenbones.nvim/issues/178)) ([a509e6c](https://github.com/mcchrish/zenbones.nvim/commit/a509e6c590a07a195cc1cb77a0afe82217a2705d))
+* support nvim 0.10 ([#176](https://github.com/mcchrish/zenbones.nvim/issues/176)) ([8877bb1](https://github.com/mcchrish/zenbones.nvim/commit/8877bb1dfc50467888ebe16aaf424e36492807c9))
+
+
+### Bug Fixes
+
+* add missing fzf-lua hl ([#179](https://github.com/mcchrish/zenbones.nvim/issues/179)) ([f74fc32](https://github.com/mcchrish/zenbones.nvim/commit/f74fc327273d627ef305f85faf72c3fd9c049c3f))
+* **ci:** turn off sync for pr ([#177](https://github.com/mcchrish/zenbones.nvim/issues/177)) ([a983937](https://github.com/mcchrish/zenbones.nvim/commit/a9839370202c62527f37f695e18df5a0576bf08e))
+* **docs:** add git mirror links ([#170](https://github.com/mcchrish/zenbones.nvim/issues/170)) ([453ec69](https://github.com/mcchrish/zenbones.nvim/commit/453ec69d82d644ee6998a3464da49d0261df9f80))
+
 ## [4.2.0](https://github.com/mcchrish/zenbones.nvim/compare/v4.1.1...v4.2.0) (2024-03-24)